feat: Add Torch Button
Add torch button to the `OSBARCScannerView`. The button should only been shown on cameras that have that capability, so it's required to move the `captureDevice` property from `OSBARCScannerViewController` to the `OSBARCScannerBehaviour` class.
Toggling the torch button implies changes to the `captureDevice`'s configuration.

References: https://outsystemsrd.atlassian.net/browse/RMET-2749