feat: Add Camera Selection
Add the `OSBARCCameraModel` that is an enumerator with all the possible values for the cameras to use: front and back.
Add an `OSBARCCameraModel` parameter to the `OSBARCManagerProtocol`'s `scanBarcode` method which is then passed into `OSBARCScannerBehaviour` in order to create the `cameraDevice` property with the selected camera. This is possible due to the `OSBARCCameraModelMappable` protocol, that allows mapping a value into the `AVCaptureDevice+Position` equivalent.

References: https://outsystemsrd.atlassian.net/browse/RMET-2754