feat: Add Scan Orientation Selection
Add a new `OSBARCOrientationModel` enumerator to deal with the possible values regarding orientation. Since this value is provided by the Cordova Wrapper, a parameter of this type needs to be added to the `OSBARCManagerProtocol`'s `scanBarcode` method and passed until the `OSBARCScannerBehaviour` for UI purposes.
In order for the Scanner view to display the correct orientation, a subclass of `UIHostingController` needs to be added and its presentation style changed to `fullScreen`. This is achieved by creating and calling the new `OSBARCScannnerViewHostingController`.
Add `OSBARCDeviceTypeModel` enumerator and `OSBARCDeviceTypeModelMappable` protocol. This is needed for the `adaptive` orientation, where the orientations to respond to depends on the type of device.
Adjust `OSBARCScannerViewController` so that the video orientation adjusts to the selected orientation, not only when opening but also after rotating the device.
Make the `OSBARCCameraModelMappable` protocol generic so that it can be used by other models. The protocol is now called `OSBARCModelMappable`.

References: https://outsystemsrd.atlassian.net/browse/RMET-2753