chore: Filter Scanning View
Set `detectBarcodeRequest`'s `regionOfInterest` property so that scanning is only applied to the specific zone for it. The frame to apply is the same as the one used to construct the `OSBARCScannerView`.
The `regionOfInterest` property is updated based on a new `scanFrameChanged` frame that is set whenever `OSBARCScannerView`'s `scanFrame` property changes value (normally, on init and device rotation).
Change the way `OSBARCScannerView` is created: instead of incrementally building the screen with the black background, have this view set to the whole screen and apply a clip for the scanning zone.
Fix the issue when `deviceOrientation` is set to `flatUp`/`flatDown`.

References: https://outsystemsrd.atlassian.net/browse/RMET-2771
References: https://outsystemsrd.atlassian.net/browse/RMET-2911